:art: feat: 전반적인 리펙터링과 예외 추가

+++ src/main/java/min/scheduleproject/schedule/service/ScheduleService.java
@@ -3,6 +3,9 @@ package min.scheduleproject.schedule.service;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import min.scheduleproject.comment.dto.response.CommentResponseDto;
+import min.scheduleproject.comment.entity.Comment;
+import min.scheduleproject.comment.repository.CommentRepository;
 import min.scheduleproject.common.exception.CustomException;
 import min.scheduleproject.common.exception.ErrorCode;
 import min.scheduleproject.schedule.dto.request.ScheduleCreateRequestDto;
@@ -13,14 +16,12 @@ import min.scheduleproject.schedule.entity.Schedule;
 import min.scheduleproject.schedule.repository.ScheduleRepository;
 import min.scheduleproject.user.entity.User;
 import min.scheduleproject.user.repository.UserRepository;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.ObjectUtils;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -31,6 +32,7 @@ import java.util.stream.Collectors;
 public class ScheduleService {
 
     private final ScheduleRepository scheduleRepository;
+    private final CommentRepository commentRepository;
     private final UserRepository userRepository;
 
     @Transactional
@@ -49,25 +51,44 @@ public class ScheduleService {
     public ScheduleGetResponseDto findScheduleByScheduleId(long scheduleId) {
         Schedule found = scheduleRepository.findById(scheduleId).orElseThrow(() -> new CustomException(ErrorCode.SCHEDULE_NOT_FOUND));
 
-        return ScheduleGetResponseDto.from(found);
+        List<Comment> comments = commentRepository.findAllBySchedule_ScheduleIdOrderByCreatedAtAsc(found.getScheduleId());
+        List<CommentResponseDto> commentResponseDtos = comments.stream().map(CommentResponseDto::of).toList();
+
+        return ScheduleGetResponseDto.from(found, commentResponseDtos);
     }
 
     public Page<ScheduleGetResponseDto> findAllSchedulesByUid(Long uid, int pageNum, int pageSize) {
         Pageable pageable = PageRequest.of(pageNum, pageSize);
 
         Page<Schedule> foundSchedules = scheduleRepository.findAllByUserUidOrderByModifiedAtDesc(pageable, uid);
-        Page<ScheduleGetResponseDto> dtos = foundSchedules.map(ScheduleGetResponseDto::from);
 
-        return dtos;
+        List<ScheduleGetResponseDto> dtos = new ArrayList<>();
+        for (Schedule schedule : foundSchedules) {
+            List<Comment> comments = commentRepository.findAllBySchedule_ScheduleIdOrderByCreatedAtAsc(schedule.getScheduleId());
+
+            List<CommentResponseDto> commentResponseDtos = comments.stream().map(CommentResponseDto::of).toList();
+
+            dtos.add(ScheduleGetResponseDto.from(schedule, commentResponseDtos));
+        }
+
+        return new PageImpl<>(dtos, pageable, foundSchedules.getTotalElements());
     }
 
     public Page<ScheduleGetResponseDto> findAllScheduleByTitle(String title, int pageNum, int pageSize) {
         Pageable pageable = PageRequest.of(pageNum, pageSize);
 
         Page<Schedule> foundSchedules = scheduleRepository.findAllByTitleOrderByModifiedAtDesc(pageable, title);
-        Page<ScheduleGetResponseDto> dtos = foundSchedules.map(ScheduleGetResponseDto::from);
 
-        return dtos;
+        List<ScheduleGetResponseDto> dtos = new ArrayList<>();
+        for (Schedule schedule : foundSchedules) {
+            List<Comment> comments = commentRepository.findAllBySchedule_ScheduleIdOrderByCreatedAtAsc(schedule.getScheduleId());
+
+            List<CommentResponseDto> commentResponseDtos = comments.stream().map(CommentResponseDto::of).toList();
+
+            dtos.add(ScheduleGetResponseDto.from(schedule, commentResponseDtos));
+        }
+
+        return new PageImpl<>(dtos, pageable, foundSchedules.getTotalElements());
     }
 
     @Transactional
@@ -88,7 +109,7 @@ public class ScheduleService {
         Long uid = (Long) session.getAttribute("LOGIN_USER");
 
         Schedule found = scheduleRepository.findById(scheduleId).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
-
+        if(commentRepository.existsBySchedule_ScheduleId(found.getScheduleId())) throw new CustomException(ErrorCode.SCHEDULE_HAS_COMMENTS);
         if(!ObjectUtils.nullSafeEquals(uid, found.getUser().getUid())) throw new CustomException(ErrorCode.USER_ID_MISMATCH);
 
         scheduleRepository.deleteById(scheduleId);
